progress: 28.79% of code matched (4329/8216 functions, 4561/6089 objects)
Source-commit: cb084d7a6f04b2541694b971dc4078a8014831b4

+++ src/game/game_fn_80129108.c
@@ -10,7 +10,7 @@ typedef struct Runtime {
 #define FN_80128E30_PARAMETERS void*
 extern FN_80128E30_RETURN fn_80128E30(FN_80128E30_PARAMETERS);
 
-s32 fn_80129108(void* owner, s32 clearMask, s32 setMask)
+s32 fn_80129108(void* owner, long clearMask, long setMask)
 {
     Runtime* runtime = fn_80128E30(owner);
     s32 oldFlags = runtime->flags;
